feat(test): in-source test for options
