+++ packages/unplugin-typia/vitest.config.js
@@ -2,6 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
 	test: {
+		globals: true,
 		includeSource: ['src/**/*.{js,ts}'],
 	},
 });
